+++ packages/eslint-config-react/index.js
@@ -3,5 +3,10 @@ const localpkg = require('local-pkg')
 const hasTS = localpkg.isPackageExists('typescript')
 
 module.exports = {
-    extends: [hasTS ? '@liting-yes/eslint-config-ts' : '@liting-yes/eslint-config-base', 'plugin:react/recommended', 'plugin:react-hooks/recommended', 'plugin:jsx-a11y/recommended']
+  'extends': [
+    hasTS ? '@liting-yes/eslint-config-ts' : '@liting-yes/eslint-config-base',
+    'plugin:react/recommended',
+    'plugin:react-hooks/recommended',
+    'plugin:jsx-a11y/recommended',
+  ],
 }