+++ packages/eslint-config-ts/index.js
@@ -1,16 +1,21 @@
 module.exports = {
-  extends: ['@liting-yes/eslint-config-base', 'plugin:@typescript-eslint/recommended', 'plugin:import/recommended', 'plugin:import/typescript'],
-  plugins: ['@typescript-eslint'],
+  'extends': [
+    '@liting-yes/eslint-config-base',
+    'plugin:@typescript-eslint/recommended',
+    'plugin:import/recommended',
+    'plugin:import/typescript',
+  ],
   overrides: [
     {
-      files: ['*.ts'],
-      parser: '@typescript-eslint/parser',
+      files:    [ '*.ts' ],
+      parser:   '@typescript-eslint/parser',
       settings: {
         'import/resolver': {
-          typescript:true,
-          node: true
-        }
-      }
+          node:       true,
+          typescript: true,
+        },
+      },
     },
   ],
+  plugins: [ '@typescript-eslint' ],
 }