+++ packages/eslint-config-vue/index.js
@@ -5,7 +5,7 @@ const isVue2 = localpkg.getPackageInfo('vue')?.version?.startsWith('2')
 
 module.exports = {
   'extends': [
-    hasTS ? 'liting-yes/eslint-config-ts' : '@liting-yes/eslint-config-base',
+    hasTS ? '@liting-yes/eslint-config-ts' : '@liting-yes/eslint-config-base',
     isVue2 ? 'plugin:vue/recommended' : 'plugin:vue/vue3-recommended',
   ],
 }