+++ packages/eslint-config-base/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @liting-yes/eslint-config-base
 
+## 0.2.2
+
+### Patch Changes
+
+- fix
+
+## 0.2.1
+
+### Patch Changes
+
+- fix
+
 ## 0.2.0
 
 ### Minor Changes